% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActionSquirrel-package.R
\docType{package}
\name{ActionSquirrel-package}
\alias{ActionSquirrel}
\alias{ActionSquirrel-package}
\title{ActionSquirrel: Generate An R6-Class 'Action Squirrel' Game}
\description{
Prints to the console a grid of emoji representing a forest,
    including a player-controlled squirrel that can be controlled by the user
    to collect nuts and avoid the owl.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matt-dray/ActionSquirrel}
  \item Report bugs at \url{https://github.com/matt-dray/ActionSquirrel/issues}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
