% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{ActionSquirrel}
\alias{ActionSquirrel}
\title{Generate An R6-Class 'Action Squirrel' Game}
\value{
Console output with \code{cat()}.
}
\description{
Generate An R6-Class 'Action Squirrel' Game

Generate An R6-Class 'Action Squirrel' Game
}
\details{
Prints to the console a grid of emoji representing an overworld,
including a player character that can be controlled by the user.
Also provides a live tally of moves taken.
}
\examples{
    \dontrun{
    x <- ActionSquirrel$new()  # generate overworld
    x$move("up")               # move character up
    x$move("right")            # move character right
    }
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{overworld}}{Character vector. The overworld in one dimension.}

\item{\code{s_loc}}{Numeric. The location of the squirrel.}

\item{\code{n_loc}}{Numeric. The location of the nut.}

\item{\code{o_loc}}{Numeric. The location of the owl.}

\item{\code{moves}}{Numeric. The number of moves made by the player.}

\item{\code{nuts}}{Numeric. The number of nuts collected by the player.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ActionSquirrel$new()}}
\item \href{#method-move}{\code{ActionSquirrel$move()}}
\item \href{#method-clone}{\code{ActionSquirrel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new overworld.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActionSquirrel$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An R6-class object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-move"></a>}}
\if{latex}{\out{\hypertarget{method-move}{}}}
\subsection{Method \code{move()}}{
Move the hero around the overworld.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActionSquirrel$move(where = c("up", "down", "left", "right"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{where}}{Character. Which direction to move. One of \code{"up"},
\code{"down"}, \code{"left"} and \code{"right"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An R6-class object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActionSquirrel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
